Allow merging schema stores
This means linter options within a configuration file correctly merge
with each other, rather than entirely overriding the previous entry.

We do this by introducing a Term.Repr.repr type, which more closely
describes the term that we parsed, in that it tracks which fields were
actually defined. This allows us to merge terms, and thus merge schemas.

Currently the term representation is entirely internal to the config
system - only exposed via Store.merge. We might need to extend this in
the future, but this'll do for now.

One downside of this, is that Schemas can no longer be converted to
Terms - one must use Schema.to_parser/Schema.write_default
directly. This isn't currently an issue, but may be more problematic in
the future.