Fix column calculations of spans
Previously the following span would be considered columns 1-4.

   foo = 0
   ^^^

I believe this is because the lexer uses the next character when
building up the positions. We want this to show as columns 1-3, and so
only add 1 to the first column.

This also requires adjusting the error printing and character escape
linter.