Rewrite how spans are stored
Previously each span stored its start/end line/col/bol. While this was
fine, it meant we didn't always have entirely complete information about
line information (such as line bols on multi-line strings, or the
ability to compute line length).

This replaces the previous storage system with the following:

 - A lookup of line positions -> BOL offsets. We can perform a binary
   search over this to find the line number for a given node.

 - Spans now store a reference to this line lookup, and a pair
   of (start, end) integers. These are used to determine line number and
   bol (and thus column).

While position lookup is unquestionably slower, this should result in
smaller memory usage (as we store three int64s, instead of 7), as well
as allowing for more information to be exposed.