Move HTML emitting options into a common type
Rather than 4 keyword arguments. There's probably better ways of doing
this, but it's nice enough for now.

Note, modules are still passed separately, as they're data which changes
across runs, while options are (notionally) constant. In practice it's
not that simple, as files in the /module/ folder have different
options (well, a different resolve:str->str) to those in the root.