Treat "documentation library path" as globs
Before they were sort of treated as Lua-style paths but not really. It's
possible we need a hybrid solution (so one can treat foo/init.lua as
foo, but this'll do for now).

Main motivation for this is to allow writing
/src/main/resources/*/computercraft/lua/rom/apis/ as the Lua path in CC:T.
Though this is somewhat pointless given that everything is in one folder
now.