Split up documentation into modules and pages
The former is a Lua module (and so usable from code), the other is just
a piece of writing. This is such a janky model, and I'm not 100% happy
with it, but it's a start I guess...

Also rename module_kind to namespace. Making the previous commit all
confusing.

---

Is it obvious that I don't know how to do system design at this point?