Vendor in omd 2
Omd 2 is much more standards compliant[1], and so it's about time to
upgrade. However, the new version is a complete rewrite and does not
provide an extensible parser, so we must vendor and modify it instead.

There are unfortunately a couple of changes in how documents are
parsed/rendered. For instance, the following:

    Foo
        local a = 0

Is now treated as a single paragraph, rather than a paragraph followed
by a code block.

This is more standards compliant, but does occur a couple of times in
CC:T's docs.

As we're vendoring our own markdown parser, I think there is some scope
for further markdown extensions. For instance:
 - Admonitions and code block extensions from Docusaurus
 - Tables and auto-links from GitHub Flavoured Markdown
 - Pandoc footnotes? Possibly more awkward to do, as we'd need to
   rewrite them to be shared across the full page.
 - Smart quotes

[1]: Still not perfect! Fixing these have been on my todo list. However,
     it's enough to fix #28.