Use lrgrep for parse errors
Largely identical to what we're doing on CC:T.

This did require making the `Stmts: list(Stmt)` production
left-recursive, meaning we need to do a `List.rev` at the end. It's a
little nasty, but ends up reducing the number of states in the parser.

Note this does rely on our own fork of lrgrep for now - hopefully will
upstream these changes, just need to clean them up first.