Rewrite the data engine to be keyed on Span.filename
This is much more complex than I'd like it to be. There's some
cases (notably linting/printing doc comments) where we want to run
analysis on a snippet of code rather than an existing file. For now, we
just expose a standalone "compute" function, but it's a little sad!

Oh goodness, I really should just nuke the LSP engine for now.