Read the path to the search index from the page's meta
We now add a <meta name="illuaminate:index" contents="xxx.json"> element
to the head of the document, storing the path to our search index. This
is then read by the embedded Javascript to load the actual index.

This ensures that the path to the index uses a relative URL, rather than
requiring it be in the root of the domain. Fixes #44.