Flatten trivia representation
Trivia is now represented as a single tuple of `(start, kind, contents)`,
rather than a nested `(Kind of contents) spanned`. We also store trivia
in an immutable array, rather than a list — mutations are not limited to
just being at the start, so we don't really benefit from a list anyway.